requirements.txt
APScheduler==3.0.0
beautifulsoup4==4.6.3
certifi==2018.8.24
chardet==3.0.4
click==6.7
cssselect==1.0.3
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.0.2
Flask-Cors==3.0.6
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.10
jsonlines==1.2.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.5
lxml==4.2.5
MarkupSafe==1.0
newspaper3k==0.2.7
nltk==3.3
numpy==1.16.0
pandas==0.23.4
Pillow==5.2.0
psycopg2==2.7.5
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
requests-file==1.4.3
rq==0.12.0
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
sqlalchemy==1.2.12
tensorflow==1.15
tinysegmenter==0.3
tldextract==2.2.0
urllib3==1.23
Werkzeug==0.14.1
